src/utils/config.py
"""Minimal YAML config loader. Returns a plain dict; callers pull the keys they need."""
from __future__ import annotations

from pathlib import Path
from typing import Any

import yaml


def load_config(path: str | Path) -> dict[str, Any]:
    """Load a YAML config file and return it as a dict."""
    path = Path(path)
    if not path.is_file():
        raise FileNotFoundError(f"Config not found: {path}")
    with path.open("r", encoding="utf-8") as f:
        cfg = yaml.safe_load(f)
    if not isinstance(cfg, dict):
        raise ValueError(f"Top-level YAML in {path} must be a mapping, got {type(cfg).__name__}")
    return cfg
