requirements.txt
numpy>=1.26,<2.2
pandas>=2.1
pyarrow>=15.0
scikit-learn>=1.4
joblib>=1.3
PyYAML>=6.0
matplotlib>=3.8

# Deep learning + RL
torch>=2.2
gymnasium>=0.29,<1.0
stable-baselines3>=2.3
tensorboard>=2.16

# MT5 data extraction (Windows-only; install only if re-pulling raw data)
# MetaTrader5>=5.0.45

# Testing & tooling
pytest>=8.0
pytest-cov>=5.0
